+++ tests/testthat/test-geom_subgene_arrow.R
@@ -5,7 +5,7 @@ test_that("a simple geom_subgene_arrow plot is drawn without errors", {
     library(ggplot2)
     p <- ggplot(
       example_genes,
-      aes(xmin = start, xmax = end, y = molecule)
+      aes(xmin = start, xmax = end, y = molecule, forward = orientation)
     ) +
       geom_gene_arrow() +
       geom_subgene_arrow(
@@ -16,13 +16,38 @@ test_that("a simple geom_subgene_arrow plot is drawn without errors", {
           y = molecule,
           fill = gene,
           xsubmin = from,
-          xsubmax = to
+          xsubmax = to,
+          forward = orientation
         )
       ) +
       facet_wrap(~ molecule, scales = "free", ncol = 1) +
       theme_genes()
     print(p)
   } )
+
+  expect_doppelganger("Basic plot with subgenes", {
+    ggplot(
+      example_genes,
+      aes(xmin = start, xmax = end, y = molecule, forward = orientation)
+    ) +
+      geom_gene_arrow() +
+      geom_subgene_arrow(
+        data = example_subgenes,
+        aes(
+          xmin = start,
+          xmax = end,
+          y = molecule,
+          fill = gene,
+          xsubmin = from,
+          xsubmax = to,
+          forward = orientation
+        )
+      ) +
+      facet_wrap(~ molecule, scales = "free", ncol = 1) +
+      theme_genes()
+
+  })
+
 })
 
 test_that("boundary breaking subgenes are caught", {