+++ R/geom_subgene_label.R
@@ -73,6 +73,7 @@ geom_subgene_label <- function(
       min.size = min.size,
       grow = grow,
       reflow = reflow,
+      height = height,
       ...
     )
   )